Move PackageStatus to scorecard/backend
