+++ app/lib/analyzer/backend.dart
@@ -10,10 +10,8 @@ import 'package:gcloud/service_scope.dart' as ss;
 import 'package:logging/logging.dart';
 import 'package:pub_semver/pub_semver.dart';
 
-import '../frontend/models.dart';
 import '../shared/analyzer_memcache.dart';
 import '../shared/analyzer_service.dart';
-import '../shared/utils.dart';
 import '../shared/versions.dart';
 
 import 'models.dart';
@@ -258,49 +256,6 @@ class AnalysisBackend {
           '${obsoleteKeys.map((k) => k.id).join(',')}');
       await db.commit(deletes: obsoleteKeys);
     }
-  }
-}
-
-class PackageStatus {
-  final bool exists;
-  final DateTime publishDate;
-  final Duration age;
-  final bool isLatestStable;
-  final bool isDiscontinued;
-  final bool isObsolete;
-  final bool isLegacy;
-  final bool usesFlutter;
-
-  PackageStatus({
-    this.exists,
-    this.publishDate,
-    this.age,
-    this.isLatestStable = false,
-    this.isDiscontinued = false,
-    this.isObsolete = false,
-    this.isLegacy = false,
-    this.usesFlutter = false,
-  });
-
-  factory PackageStatus.fromModels(Package p, PackageVersion pv) {
-    if (p == null || pv == null) {
-      return new PackageStatus(exists: false);
-    }
-    final publishDate = pv.created;
-    final isLatestStable = p.latestVersion == pv.version;
-    final now = new DateTime.now().toUtc();
-    final age = now.difference(publishDate).abs();
-    final isObsolete = age > twoYears && !isLatestStable;
-    return new PackageStatus(
-      exists: true,
-      publishDate: publishDate,
-      age: age,
-      isLatestStable: isLatestStable,
-      isDiscontinued: p.isDiscontinued ?? false,
-      isObsolete: isObsolete,
-      isLegacy: pv.pubspec.supportsOnlyLegacySdk,
-      usesFlutter: pv.pubspec.usesFlutter,
-    );
   }
 }
